+++ src/components/signup/SignupForm.js
@@ -1,176 +1,93 @@
-import React, { useState } from "react";
-import { makeStyles } from "@material-ui/core/styles";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import AppBar from "@material-ui/core/AppBar";
-import Toolbar from "@material-ui/core/Toolbar";
+import React, { Component } from "react";
 import Paper from "@material-ui/core/Paper";
 import Stepper from "@material-ui/core/Stepper";
 import Step from "@material-ui/core/Step";
 import StepLabel from "@material-ui/core/StepLabel";
 import Button from "@material-ui/core/Button";
-import Link from "@material-ui/core/Link";
 import Typography from "@material-ui/core/Typography";
 
 import SignupDetails from "./SignupDetails";
 import AddressForm from "./AddressForm";
-// import PaymentForm from "./PaymentForm";
-// import Review from "./Review";
 
-const useStyles = makeStyles(theme => ({
-  appBar: {
-    position: "relative"
-  },
-  layout: {
-    width: "auto",
-    marginLeft: theme.spacing(2),
-    marginRight: theme.spacing(2),
-    [theme.breakpoints.up(600 + theme.spacing(2) * 2)]: {
-      width: 600,
-      marginLeft: "auto",
-      marginRight: "auto"
+export default class SignupForm extends Component {
+  state = {
+    activeStep: 0,
+    firstName: "jend",
+    lastName: "hordejan",
+    email: "hello",
+    password: "pass",
+    confirmPassword: "pass",
+    imageUrl:
+      "https://media.istockphoto.com/vectors/colorful-brown-circle-chef-logo-vector-id1056400912",
+    houseNo: "",
+    postCode: "",
+    about: "",
+    verifiedAddress: "",
+    address: ""
+  };
+
+  render() {
+    const steps = [
+      "SignIn details",
+      "Personal details",
+      "Review your registration"
+    ];
+
+    function getStepContent(step, handleChange, values) {
+      switch (step) {
+        case 0:
+          return <SignupDetails handleChange={handleChange} values={values} />;
+
+        case 1:
+          //   return <PaymentForm />;
+          return (
+            <AddressForm
+              handleChange={handleChange}
+              values={values}
+              handleImageUrlChange={handleImageUrlChange}
+            />
+          );
+        case 2:
+          //   return <Review />;
+          return <AddressForm />;
+        default:
+          throw new Error("Unknown step");
+      }
     }
-  },
-  paper: {
-    marginTop: theme.spacing(3),
-    marginBottom: theme.spacing(3),
-    padding: theme.spacing(2),
-    [theme.breakpoints.up(600 + theme.spacing(3) * 2)]: {
-      marginTop: theme.spacing(6),
-      marginBottom: theme.spacing(6),
-      padding: theme.spacing(3)
-    }
-  },
-  stepper: {
-    padding: theme.spacing(3, 0, 5)
-  },
-  buttons: {
-    display: "flex",
-    justifyContent: "flex-end"
-  },
-  button: {
-    marginTop: theme.spacing(3),
-    marginLeft: theme.spacing(1)
-  }
-}));
 
-const steps = [
-  "SignIn details",
-  "Personal details",
-  "Review your registration"
-];
+    const handleImageUrlChange = value => {
+      console.log("Lets finally set the IMAGEURL:", value);
 
-function getStepContent(step, handleChange, values, handleImageUrlChange) {
-  switch (step) {
-    case 0:
-      return <SignupDetails handleChange={handleChange} values={values} />;
-
-    case 1:
-      //   return <PaymentForm />;
-      return (
-        <AddressForm
-          handleChange={handleChange}
-          values={values}
-          handleImageUrlChange={handleImageUrlChange}
-        />
+      this.setState({ imageUrl: value });
+      console.log(
+        "let's check imageURL from SignupForm: ",
+        this.state.imageUrl
       );
-    case 2:
-      //   return <Review />;
-      return <AddressForm />;
-    default:
-      throw new Error("Unknown step");
-  }
-}
+    };
 
-export default function SignupForm() {
-  const [image, setImage] = useState("");
-  const [firstName, setFirstName] = useState("J");
-  const [lastName, setLastName] = useState("H");
-  const [email, setEmail] = useState("j@gmail");
-  const [password, setPassword] = useState("pass");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [imageUrl, setImageUrl] = useState("");
-  const [houseNo, setHouseNo] = useState("");
-  const [postCode, setPostCode] = useState("");
-  const [about, setAbout] = useState("");
-  const [verifiedAddress, setVerifiedAddress] = useState(false);
-  const [address, setAddress] = useState(null);
+    const handleChange = input => e => {
+      this.setState({ [input]: e.target.value });
+    };
 
-  const values = {
-    firstName,
-    lastName,
-    email,
-    password,
-    confirmPassword,
-    imageUrl,
-    houseNo,
-    postCode,
-    about,
-    verifiedAddress,
-    address
-  };
+    const handleNext = () => {
+      this.setState({ activeStep: this.state.activeStep + 1 });
+    };
 
-  const handleChange = input => e => {
-    // this.setState({ [input]: e.target.value });
-    // setImage(file.secure_url);
-    console.log(`input: ${input} | value ${e.target.value}`);
+    const handleBack = () => {
+      this.setState({ activeStep: this.state.activeStep - 1 });
+    };
 
-    switch (input) {
-      case "firstName": {
-        setFirstName(e.target.value);
-        return;
-      }
-      case "lastName": {
-        setLastName(e.target.value);
-        return;
-      }
-      case "email": {
-        setEmail(e.target.value);
-        return;
-      }
-      case "password": {
-        setPassword(e.target.value);
-        return;
-      }
-      case "confirmPassword": {
-        setConfirmPassword(e.target.value);
-        return;
-      }
-      case "imageUrl": {
-        setImageUrl(e.target.value);
-        console.log("new value of imageUrl:", confirmPassword);
-        return;
-      }
-    }
-  };
-
-  const handleImageUrlChange = value => {
-    console.log("Lets finally set the IMAGEURL:", value);
-
-    setFirstName(value);
-    console.log("let's check imageURL from SignupForm: ", firstName);
-  };
-
-  const classes = useStyles();
-  const [activeStep, setActiveStep] = React.useState(0);
-
-  const handleNext = () => {
-    setActiveStep(activeStep + 1);
-  };
-
-  const handleBack = () => {
-    setActiveStep(activeStep - 1);
-  };
-
-  return (
-    <React.Fragment>
-      <CssBaseline />
-
+    const { classes } = this.props;
+    return (
       <main className={classes.layout}>
         <Paper className={classes.paper}>
           <Typography component="h1" variant="h4" align="center">
             Registration
           </Typography>
-          <Stepper activeStep={activeStep} className={classes.stepper}>
+          <Stepper
+            activeStep={this.state.activeStep}
+            className={classes.stepper}
+          >
             {steps.map(label => (
               <Step key={label}>
                 <StepLabel>{label}</StepLabel>
@@ -178,7 +95,7 @@ export default function SignupForm() {
             ))}
           </Stepper>
           <React.Fragment>
-            {activeStep === steps.length ? (
+            {this.state.activeStep === steps.length ? (
               <React.Fragment>
                 <Typography variant="h5" gutterBottom>
                   Thank you for your order.
@@ -192,13 +109,12 @@ export default function SignupForm() {
             ) : (
               <React.Fragment>
                 {getStepContent(
-                  activeStep,
+                  this.state.activeStep,
                   handleChange,
-                  values,
-                  handleImageUrlChange
+                  this.state
                 )}
                 <div className={classes.buttons}>
-                  {activeStep !== 0 && (
+                  {this.state.activeStep !== 0 && (
                     <Button onClick={handleBack} className={classes.button}>
                       Back
                     </Button>
@@ -209,7 +125,9 @@ export default function SignupForm() {
                     onClick={handleNext}
                     className={classes.button}
                   >
-                    {activeStep === steps.length - 1 ? "Place order" : "Next"}
+                    {this.state.activeStep === steps.length - 1
+                      ? "Place order"
+                      : "Next"}
                   </Button>
                 </div>
               </React.Fragment>
@@ -217,6 +135,6 @@ export default function SignupForm() {
           </React.Fragment>
         </Paper>
       </main>
-    </React.Fragment>
-  );
+    );
+  }
 }
